denoised this to only talk about deleting when non-zero

+++ djasana/management/commands/sync_from_asana.py
@@ -276,11 +276,12 @@ class Command(BaseCommand):
             # Delete local tasks for this project that are no longer in Asana.
             tasks_to_delete = Task.objects.filter(projects=project).exclude(
                 remote_id__in=remote_task_ids).exclude(remote_id__isnull=True)
-            id_list = list(tasks_to_delete.values_list('remote_id', flat=True))
-            message = "Deleting Tasks No Longer Present " + str(id_list)
-            self.stdout.write(self.style.SUCCESS(message))
-            logger.info(message)
-            tasks_to_delete.delete()
+            if tasks_to_delete.count() > 0:
+                id_list = list(tasks_to_delete.values_list('remote_id', flat=True))
+                tasks_to_delete.delete()
+                message = "Deleted {} Tasks No Longer Present: {} ".format(len(id_list), str(id_list))
+                self.stdout.write(self.style.SUCCESS(message))
+                logger.info(message)
         if project:
             message = 'Successfully synced project {}.'.format(project.name)
             self.stdout.write(self.style.SUCCESS(message))
